refactor(85): split EndgameOverallPerformanceSection into per-card files
Break the 802-LOC composite section into five sibling files under
frontend/src/components/charts/ to comply with the 200-LOC hard limit:

- EndgameOverallCard.tsx          — Cards 1 + 3 (WDL + sig-gated score)
- EndgameOverallEntryCard.tsx     — Card 2 (entry eval + achievable score)
- EndgameOverallScoreGapRow.tsx   — Gap row (label + math + bullet)
- EndgameOverallConnectorArrows.tsx — Desktop arrow overlay + DOM geometry
- EndgameOverallShared.ts         — deriveLevel + domain constants

The orchestrator file now holds only the per-card score math, the
operandColor / gapZoneColor helpers, and the grid JSX. No behavior change;
all 346 frontend tests pass, lint and typecheck clean.
